+++ src/agora/radio/radio_set/radio_set_bs.h
@@ -1,11 +1,13 @@
 /**
- * @file radio_lib_uhd.h
- * @brief Declaration file for the BsRadioSet class.
+ * @file radio_set_bs.h
+ * @brief Declaration file for the RadioSetBs class.
  */
-#ifndef RADIO_LIB_UHD_H_
-#define RADIO_LIB_UHD_H_
+#ifndef RADIO_SET_BS_H_
+#define RADIO_SET_BS_H_
 
+#include <atomic>
 #include <complex>
+#include <cstdint>
 #include <cstdlib>
 #include <memory>
 #include <vector>
@@ -15,43 +17,29 @@
 #include "config.h"
 #include "memory_manage.h"
 #include "radio.h"
-#include "uhd/usrp/multi_usrp.hpp"
+#include "radio_set.h"
 
-class RadioUHDConfig {
+class RadioSetBs : public RadioSet {
  public:
-  RadioUHDConfig(Config* cfg, Radio::RadioType radio_type);
-  ~RadioUHDConfig();
+  RadioSetBs(Config* cfg, Radio::RadioType radio_type);
+  virtual ~RadioSetBs() final;
 
-  bool RadioStart();
-  void RadioStop();
-  void ReadSensors();
-  int RadioTx(size_t radio_id, const void* const* buffs, Radio::TxFlags flags,
-              long long& tx_time);
-  int RadioTx(size_t radio_id,
-              const std::vector<std::vector<std::complex<int16_t>>>& tx_data,
-              Radio::TxFlags flags, long long& tx_time_ns);
+  virtual bool RadioStart() final;
+  virtual void Go() final;
 
-  int RadioRx(size_t radio_id,
-              std::vector<std::vector<std::complex<int16_t>>>& rx_data,
-              size_t rx_size, Radio::RxFlags& out_flags, long long& rx_time_ns);
-
-  int RadioRx(size_t radio_id,
-              std::vector<std::vector<std::complex<int16_t>>*>& rx_buffs,
-              size_t rx_size, Radio::RxFlags& out_flags, long long& rx_time_ns);
-
-  int RadioRx(size_t radio_id, std::vector<void*>& rx_locs, size_t rx_size,
-              Radio::RxFlags& out_flags, long long& rx_time_ns);
-
-  bool DoCalib() const { return calib_; }
-  void Go();
-  arma::cx_float* GetCalibUl() { return init_calib_ul_processed_; }
-  arma::cx_float* GetCalibDl() { return init_calib_dl_processed_; }
-
-  // Thread functions
-  void InitBsRadio(size_t radio_id);
-  void ConfigureBsRadio(size_t radio_id);
+  virtual bool DoCalib() const final { return calib_; }
+  virtual arma::cx_float* GetCalibUl() final {
+    return init_calib_ul_processed_;
+  }
+  virtual arma::cx_float* GetCalibDl() final {
+    return init_calib_dl_processed_;
+  }
 
  private:
+  // Thread functions
+  void InitRadio(size_t radio_id);
+  void ConfigureRadio(size_t radio_id);
+
   long long SyncArrayTime();
 
   void CalibrateSampleOffset();
@@ -73,15 +61,12 @@ class RadioUHDConfig {
                            double tx_center_tone);
   static void SetIqBalance(Radio* dev, int direction, size_t channel, int gcorr,
                            int iqcorr);
-
-  // TODO: should not be a vector, instead should be a single object; Modify when testing
   static void AdjustCalibrationGains(std::vector<Radio*>& rx_devs,
                                      Radio* tx_dev, size_t channel,
                                      double fft_bin, bool plot = false);
   void DciqCalibrationProc(size_t channel);
   Config* cfg_;
-  uhd::usrp::multi_usrp::sptr hubs_;
-  std::unique_ptr<Radio> radios_;
+  std::vector<SoapySDR::Device*> hubs_;
   arma::cx_float* init_calib_ul_processed_;
   arma::cx_float* init_calib_dl_processed_;
   Table<arma::cx_float> init_calib_ul_;
@@ -94,4 +79,4 @@ class RadioUHDConfig {
   std::atomic<size_t> num_radios_initialized_;
   std::atomic<size_t> num_radios_configured_;
 };
-#endif  // RADIO_LIB_UHD_H_
+#endif  // RADIO_SET_BS_H_
