+++ src/agora/radio/radio_set/radio_set_uhd.h
@@ -1,13 +1,11 @@
 /**
- * @file bs_radio_set.h
- * @brief Declaration file for the BsRadioSet class.
+ * @file radio_set_uhd.h
+ * @brief Declaration file for the RadioSetUhd class.
  */
-#ifndef BS_RADIO_SET_H_
-#define BS_RADIO_SET_H_
+#ifndef RADIO_SET_UHD_H_
+#define RADIO_SET_UHD_H_
 
-#include <atomic>
 #include <complex>
-#include <cstdint>
 #include <cstdlib>
 #include <memory>
 #include <vector>
@@ -17,29 +15,43 @@
 #include "config.h"
 #include "memory_manage.h"
 #include "radio.h"
-#include "radio_set.h"
+#include "uhd/usrp/multi_usrp.hpp"
 
-class BsRadioSet : public RadioSet {
+class RadioSetUhd {
  public:
-  BsRadioSet(Config* cfg, Radio::RadioType radio_type);
-  virtual ~BsRadioSet() final;
+  RadioSetUhd(Config* cfg, Radio::RadioType radio_type);
+  ~RadioSetUhd();
 
-  virtual bool RadioStart() final;
-  virtual void Go() final;
+  bool RadioStart();
+  void RadioStop();
+  void ReadSensors();
+  int RadioTx(size_t radio_id, const void* const* buffs, Radio::TxFlags flags,
+              long long& tx_time);
+  int RadioTx(size_t radio_id,
+              const std::vector<std::vector<std::complex<int16_t>>>& tx_data,
+              Radio::TxFlags flags, long long& tx_time_ns);
 
-  virtual bool DoCalib() const final { return calib_; }
-  virtual arma::cx_float* GetCalibUl() final {
-    return init_calib_ul_processed_;
-  }
-  virtual arma::cx_float* GetCalibDl() final {
-    return init_calib_dl_processed_;
-  }
+  int RadioRx(size_t radio_id,
+              std::vector<std::vector<std::complex<int16_t>>>& rx_data,
+              size_t rx_size, Radio::RxFlags& out_flags, long long& rx_time_ns);
+
+  int RadioRx(size_t radio_id,
+              std::vector<std::vector<std::complex<int16_t>>*>& rx_buffs,
+              size_t rx_size, Radio::RxFlags& out_flags, long long& rx_time_ns);
+
+  int RadioRx(size_t radio_id, std::vector<void*>& rx_locs, size_t rx_size,
+              Radio::RxFlags& out_flags, long long& rx_time_ns);
+
+  bool DoCalib() const { return calib_; }
+  void Go();
+  arma::cx_float* GetCalibUl() { return init_calib_ul_processed_; }
+  arma::cx_float* GetCalibDl() { return init_calib_dl_processed_; }
+
+  // Thread functions
+  void InitBsRadio(size_t radio_id);
+  void ConfigureBsRadio(size_t radio_id);
 
  private:
-  // Thread functions
-  void InitRadio(size_t radio_id);
-  void ConfigureRadio(size_t radio_id);
-
   long long SyncArrayTime();
 
   void CalibrateSampleOffset();
@@ -61,12 +73,15 @@ class BsRadioSet : public RadioSet {
                            double tx_center_tone);
   static void SetIqBalance(Radio* dev, int direction, size_t channel, int gcorr,
                            int iqcorr);
+
+  // TODO: should not be a vector, instead should be a single object; Modify when testing
   static void AdjustCalibrationGains(std::vector<Radio*>& rx_devs,
                                      Radio* tx_dev, size_t channel,
                                      double fft_bin, bool plot = false);
   void DciqCalibrationProc(size_t channel);
   Config* cfg_;
-  std::vector<SoapySDR::Device*> hubs_;
+  uhd::usrp::multi_usrp::sptr hubs_;
+  std::unique_ptr<Radio> radios_;
   arma::cx_float* init_calib_ul_processed_;
   arma::cx_float* init_calib_dl_processed_;
   Table<arma::cx_float> init_calib_ul_;
@@ -79,4 +94,4 @@ class BsRadioSet : public RadioSet {
   std::atomic<size_t> num_radios_initialized_;
   std::atomic<size_t> num_radios_configured_;
 };
-#endif  // BS_RADIO_SET_H_
+#endif  // RADIO_SET_UHD_H_
