+++ src/agora/txrx/workers/txrx_worker_usrp.cc
@@ -24,7 +24,7 @@ TxRxWorkerUsrp::TxRxWorkerUsrp(
     moodycamel::ProducerToken& notify_producer,
     std::vector<RxPacket>& rx_memory, std::byte* const tx_memory,
     std::mutex& sync_mutex, std::condition_variable& sync_cond,
-    std::atomic<bool>& can_proceed, RadioUHDConfig& radio_config)
+    std::atomic<bool>& can_proceed, RadioSetUhd& radio_config)
     : TxRxWorker(core_offset, tid, radio_hi, radio_lo, config->NumChannels(),
                  config, rx_frame_start, event_notify_q, tx_pending_q,
                  tx_producer, notify_producer, rx_memory, tx_memory, sync_mutex,
